% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FastGibbsSamplers-package.R
\docType{package}
\name{FastGibbsSamplers}
\alias{FastGibbsSamplers-package}
\alias{FastGibbsSamplers}
\title{Fast Gibbs Samplers: Blocked, Partially Collapsed, and Nested Gibbs Samplers}
\description{
Provides fast and scalable Gibbs sampling algorithms for
Bayesian penalized linear regression models in high-dimensional
settings. The package implements efficient partially collapsed
and nested Gibbs samplers for Bayesian Lasso, with a focus on
computational efficiency when the number of predictors is large
relative to the sample size.
}
\author{
\strong{Maintainer}: Mohammad Javad Davoudabadi \email{mohammad.davoudabadi@qut.edu.au} [copyright holder]

Authors:
\itemize{
  \item John Ormerod \email{john.ormerod@sydney.edu.au} (\href{https://orcid.org/0000-0002-4650-7507}{ORCID})
  \item Garth Tarr \email{garth.tarr@gmail.com} (\href{https://orcid.org/0000-0002-6605-7478}{ORCID})
  \item Samuel Mueller \email{samuel.mueller@sydney.edu.au} (\href{https://orcid.org/0000-0002-3087-8127}{ORCID})
}

}
